Bug #1552589  bsp: sc2332: pclint warning
[root cause ] pclint warning
[changes ] pclint warning
[side effects] pclint warning
[self test ]:pass
[change_type] pclint warning
[download normally]: YES
[power on/off normally]:YES
[do common repository/branch inspection]:NA
[is there dependence]:no dependence
[confirm dependent commit]:no dependent
[board]:ums512
Test: add report ip addr log
[test case]: add report ip addr log
[reviewers ] keguang.zhang hongwei.cao beolei.yuan

[change_type ] feature
[tag_product ] common

+++ wlan/sc2332/vendor.c
@@ -1895,7 +1895,7 @@ static int sprdwl_parse_sae_entry(struct net_device *ndev, struct sprdwl_sae_ent
 			nla_memcpy(entry->peer_addr, pos, ETH_ALEN);
 			break;
 		case SPRDWL_VENDOR_SAE_VLAN_ID:
-			entry->vlan_id = nla_get_u32(pos);
+			entry->vlan_id = (s32)nla_get_u32(pos);
 			break;
 		default:
 			break;
